requirements.txt
# Polymarket SDK & Trading
py-clob-client>=0.5.0
web3>=6.0.0

# Async I/O
websockets>=12.0
aiohttp>=3.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI & Logging
rich>=13.0.0

# Utilities
dataclasses-json>=0.6.0
